requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.28.1
pydantic==2.11.3
a2a-sdk==0.3.3

